requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Recommender Systems
scikit-surprise==1.1.3
implicit==0.7.0

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.0

# Web Framework
flask==2.3.3
flask-wtf==1.1.1
wtforms==3.0.1

# Database
sqlalchemy==2.0.23
dataset==1.6.0

# Utilities
tqdm==4.66.1
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2

# API (for future expansion)
requests==2.31.0